requirements.txt
pycryptodome==3.14.1
setproctitle==1.2.3
config==0.5.1
